9/12 commit

+++ src/main/java/com/example/sportzone/Repository/FactureRepository.java
@@ -3,5 +3,5 @@ package com.example.sportzone.Repository;
 import com.example.sportzone.entity.Facture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FactureRepository extends JpaRepository<Facture,Integer> {
+public interface FactureRepository extends JpaRepository<Facture,Long> {
 }
