+++ src/main/java/com/example/sportzone/Repository/ProprietairesalleRepository.java
@@ -3,5 +3,5 @@ package com.example.sportzone.Repository;
 import com.example.sportzone.entity.Proprietairesalle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProprietairesalleRepository extends JpaRepository<Proprietairesalle,Integer> {
+public interface ProprietairesalleRepository extends JpaRepository<Proprietairesalle,Long> {
 }
