+++ src/main/java/com/example/sportzone/Repository/SalledesportRepository.java
@@ -3,5 +3,5 @@ package com.example.sportzone.Repository;
 import com.example.sportzone.entity.Salledesport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalledesportRepository extends JpaRepository<Salledesport,Integer> {
+public interface SalledesportRepository extends JpaRepository<Salledesport,Long> {
 }
