+++ src/main/java/com/example/sportzone/Repository/ServiceRepository.java
@@ -3,5 +3,5 @@ package com.example.sportzone.Repository;
 import com.example.sportzone.entity.ServiceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<ServiceEntity,Integer> {
+public interface ServiceRepository extends JpaRepository<ServiceEntity,Long> {
 }
